Add sha512sum check to build
Since the download selects a random mirror, we should probably verify
that the downloaded archive is what it claims to be.

Can't use the gpg-generated digests provided by Kafka directly, because
those can't be passed to any tool for verification (unsurprisingly, gpg
requires you to eyeball it). So, we'll have to take the shasum, verify
it by eye, and then stick it in a build arg.

The Kafka maintainers did something very strange by publishing
a sha512sum that can't be verified with existing tools.